Minimize flake
There was a lot in here attempting to make building and running
Erlang/OTP and/or Rabbit and other tools possible within NixOS. Bazel
doesn't play nicely with NixOS so I gave up on that in favor of using
distrobox for any building, even when it's possible to build (e.g. with
rebar).

So let's minimize the flake for what I use it for in the end: perf-test
and perf. The `mkFlamegraph` script was cool but I prefer looking at
flamegraphs in Hotspot now.